3. Set canvas in world space. Configured scaling parameters to allow for coordinate system. Adjusted camera to fit canvas
For a game like GlitchGarden, we want the Unity to render the canvas in the world space. This is because game objects are going to be UI elements (eg clicking on an ally to upgrade him)

Setting up Canvas for coordination
- 1 grass square = 1 world unit
- Our canvas width is 1920 pixels
- one square is 160x160
- Number of squares = 1920/160 = 12 (therefore, we want 12 world units as our length)
- So we scale (shrink) the canvas by 12/1920 = 0.00625
- We change our canvas's X and Y scale to 0.00625 (no need for Z since this game is 2d)
- Moved Canvas from (0,0) to (5,3) - the grid coordinates now act as an array.
(try changing the postion of the coordinate checker to (1,1) (1,2) (1,3) to better understand )

As for the camera fitting the canvas, since the background height is 1080, the camera size was set to 1080 / 160 / 2 =  3.375. I positioned it to (5,3) which is the same as the canvas

Lastly, added a layer of trees for asthetics